Updates siren lights manager variables
Updates the types of variables storing time values in the siren lights manager to ensure compatibility with the `millis()` return type.

This change prevents potential issues arising from implicit type conversions.

+++ common/PAC-driver/g_errors.cpp
@@ -581,8 +581,8 @@ void siren_lights_manager::eval()
     if ( tech_obj_error::is_any_message )
         {
         static int step = 0;
-        static unsigned long start_blink_time = 0;
-        static unsigned long start_wait_time = 0;
+        static uint32_t start_blink_time = 0;
+        static uint32_t start_wait_time = 0;
 
         switch ( step )
             {
